refactor: move nginx related functions to client_utils
